Remove LoadingState
Remove stale references to it too.
